+++ lookup_minifig_bricklink.py
@@ -6,7 +6,7 @@ import html
 import time
 import random
 
-import libbrick
+import libbrick.common
 import libbrick.path_utils
 import libbrick.wrappers.bricklink_wrapper as bricklink_wrapper
 
@@ -21,9 +21,9 @@ if __name__ == '__main__':
 		print("usage: ./lookupLego.py <csv txt file with lego IDs>")
 		sys.exit(1)
 
-	minifigIDpairs = libbrick.read_minifigIDpairs_from_file(minifigIDFile)
+	minifigIDpairs = libbrick.common.read_minifigIDpairs_from_file(minifigIDFile)
 
-	timestamp = libbrick.make_timestamp()
+	timestamp = libbrick.common.make_timestamp()
 	output_dir = libbrick.path_utils.get_output_dir(subdir='lookup')
 	csvfile = os.path.join(output_dir, "minifig_data-bricklink-{0}.csv".format(timestamp))